function that convert a string to an integer.

+++ 0x05-pointers_arrays_strings/100-atoi.c
@@ -6,7 +6,7 @@
   */
 int _atoi(char *s)
 {
-	 int result = 0;
+	float result = 0;
 	int sign = 1;
 	int found = 0;
 	int min = 0;
